Task3/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
bcrypt==5.0.0
click==8.3.2
colorama==0.4.6
Deprecated==1.3.1
fastapi==0.136.0
h11==0.16.0
idna==3.11
limits==5.8.0
packaging==26.1
passlib==1.7.4
pydantic==2.13.3
pydantic_core==2.46.3
PyJWT==2.12.1
python-dotenv==1.2.2
slowapi==0.1.9
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.44.0
wrapt==2.1.2
